+++ S2VX/CameraMoveCommand.hpp
@@ -4,7 +4,7 @@
 namespace S2VX {
 	class CameraMoveCommand : public CameraCommand {
 	public:
-		explicit CameraMoveCommand(Camera& camera, const int start, const int end, const EasingType easing, const int startX, const int startY, const int endX, const int endY);
+		explicit CameraMoveCommand(Camera& camera, const int start, const int end, const EasingType easing, const float startX, const float startY, const float endX, const float endY);
 		void update(const float easing);
 		const glm::vec2& getEndCoordinate() const { return endCoordinate; }
 		const glm::vec2& getStartCoordinate() const { return startCoordinate; }
