+++ S2VX/SpriteMoveCommand.cpp
@@ -1,7 +1,7 @@
 #include "SpriteMoveCommand.hpp"
 #include "Sprite.hpp"
 namespace S2VX {
-	SpriteMoveCommand::SpriteMoveCommand(Sprite* const sprite, const int start, const int end, const EasingType easing, const int pStartX, const int pStartY, const int pEndX, const int pEndY)
+	SpriteMoveCommand::SpriteMoveCommand(Sprite* const sprite, const int start, const int end, const EasingType easing, const float pStartX, const float pStartY, const float pEndX, const float pEndY)
 		: SpriteCommand{ sprite, start, end, easing },
 		startCoordinate{ pStartX, pStartY },
 		endCoordinate{ pEndX, pEndY } {}
