+++ S2VX/SpriteMoveCommand.hpp
@@ -4,7 +4,7 @@
 namespace S2VX {
 	class SpriteMoveCommand : public SpriteCommand {
 	public:
-		explicit SpriteMoveCommand(Sprite* const sprite, const int start, const int end, const EasingType easing, const int pStartX, const int pStartY, const int pEndX, const int pEndY);
+		explicit SpriteMoveCommand(Sprite* const sprite, const int start, const int end, const EasingType easing, const float pStartX, const float pStartY, const float pEndX, const float pEndY);
 		void update(const float easing);
 		const glm::vec2& getEndCoordinate() const { return endCoordinate; }
 		const glm::vec2& getStartCoordinate() const { return startCoordinate; }
